requirements.txt
google-cloud-pubsub
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
google-cloud-api-keys